components to snippets
